feat: 로그아웃 버튼 드롭다운에 구현
